requirements.txt
# Automatically generated by https://github.com/damnever/pigar.
redis==5.0.1
uvicorn==0.27.1
python-dotenv==1.00
aiohttp==3.8.4
celery==5.3.6
celery[redis]==5.3.6
fastapi==0.104.0
lmdb==1.4.1
lmdbm==0.0.5
multidict==6.0.4
pydantic==2.4.2
selectolax==0.3.20
tldextract==3.4.1
ultimate-sitemap-parser==0.5
xxhash==3.4.1
pytest==8.0.0